Delete Functionality // Error Handle
In CD frontend use command 'npm axios install' used to fetch api calls easier
----------------------------------
Theres a lot of error handling left over Im going to leave in because i was having a headache with it and I want to not let it go to waste :p. Also updated the history page react to have individual keys for the objects so it supposedly loads better according to a warning it was throwing in VSCode.