CSS changes on Create page
Changed the css on create recipe page.

-----
made a card for the text boxes. When hovering over the box yellow shaddow swings to the left side.

----
once we have a set history page design, I want to post the photo next to the create card to give users an example.