Edit Function
Added Edit route towards the backend, also added new import to use logging because I was having a lot of issues with saving from the frontend.

Additionally, added button and save button for edit function. It causes a lot of updates for the console logs on insepct, lmk if we need to get rid of that.